software prototyping/requirements.txt
asyncio==3.4.3
dbTable==0.0.4
docopt==0.6.2
git-clone==1.0.6
hbmqtt==0.9.6
numpy==1.20.2
paho-mqtt==1.5.1
passlib==1.7.4
PyYAML==5.4.1
six==1.15.0
tabulate==0.8.9
transitions==0.8.8
websockets==8.1
